Rename connected_codel to adjacent_codel
